+++ mr/src/main/java/org/opensearch/hadoop/cfg/ConfigurationOptions.java
@@ -75,6 +75,10 @@ public interface ConfigurationOptions {
     String OPENSEARCH_NODES_WAN_ONLY = "opensearch.nodes.wan.only";
     String OPENSEARCH_NODES_WAN_ONLY_DEFAULT = "false";
 
+    /** Serverless mode */
+    String OPENSEARCH_SERVERLESS = "opensearch.serverless";
+    String OPENSEARCH_SERVERLESS_DEFAULT = "false";
+
     String OPENSEARCH_NODES_RESOLVE_HOST_NAME = "opensearch.nodes.resolve.hostname";
 
     /** Secure Settings Keystore */
@@ -125,6 +129,10 @@ public interface ConfigurationOptions {
     /** Scroll keep-alive */
     String OPENSEARCH_SCROLL_KEEPALIVE = "opensearch.scroll.keepalive";
     String OPENSEARCH_SCROLL_KEEPALIVE_DEFAULT = "5m";
+
+    /** PIT keep-alive (serverless mode) */
+    String OPENSEARCH_PIT_KEEPALIVE = "opensearch.pit.keepalive";
+    String OPENSEARCH_PIT_KEEPALIVE_DEFAULT = "5m";
 
     /** Scroll size */
     String OPENSEARCH_SCROLL_SIZE = "opensearch.scroll.size";
@@ -345,4 +353,5 @@ public interface ConfigurationOptions {
 
     String OPENSEARCH_AWS_SIGV4_SERVICE_NAME = "opensearch.aws.sigv4.service.name";
     String OPENSEARCH_AWS_SIGV4_SERVICE_NAME_DEFAULT = "es";
+    String OPENSEARCH_AWS_SIGV4_SERVICE_NAME_SERVERLESS = "aoss";
 }